@@ -14,7 +14,7 @@ import {
   Button,
 } from "@chakra-ui/react";
 import { collection, getDocs, limit, orderBy, query } from "firebase/firestore";
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
 import React, { useCallback, useEffect, useState } from "react";
 import { IoPeopleCircleOutline } from "react-icons/io5";
 
