@@ -15,7 +15,7 @@ import {
 import { getDocs } from "firebase/firestore";
 import { deleteObject, ref } from "firebase/storage";
 import { useAtom, useAtomValue, useSetAtom } from "jotai";
-import { Router, useRouter } from "next/router";
+import { useRouter } from "next/navigation";
 import React, { useEffect } from "react";
 import { useAuthState } from "react-firebase-hooks/auth";
 import useCustomToast from "./useCustomToast";
