@@ -1,3 +1,5 @@
+"use client";
+
 import { authModalStateAtom } from "@/atoms/authModalAtom";
 import About from "@/components/Community/About";
 import PageContent from "@/components/Layout/PageContent";
@@ -10,17 +12,8 @@ import { useSetAtom } from "jotai";
 import React from "react";
 import { useAuthState } from "react-firebase-hooks/auth";
 
-/**
- * Post submission page where the user can create a new post.
- * If the user is not logged in, they will be prompted to log in.
- * Displays:
- * - Post creation form
- * - Community information card
- * @returns {React.FC} - Submit post page component
- */
 const SubmitPostPage: React.FC = () => {
   const [user] = useAuthState(auth);
-  // const communityStateValue = useRecoilValue(communityState);
   const { communityStateValue } = useCommunityData();
   const setAuthModalState = useSetAtom(authModalStateAtom);
 
