@@ -1,4 +1,7 @@
-import { Button, Flex, Link, Stack, Text } from "@chakra-ui/react";
+"use client";
+
+import { Button, Flex, Stack, Text } from "@chakra-ui/react";
+import Link from "next/link";
 import React from "react";
 
 const PageNotFound: React.FC = () => {
